pyproject.toml
[project]
name = "venice-arbitrage-monitor"
version = "0.1.0"
description = "Automated monitoring and alerting for VVV/DIEM token arbitrage opportunities"
authors = [
    {name = "Roderick Ford", email = "rford@swdatasci.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "web3>=6.15.0",
    "aiohttp>=3.9.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "asyncio>=3.4.3",
    "ta>=0.11.0",  # Technical analysis indicators
    "pycoingecko>=3.1.0",
    "discord-webhook>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
